backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.2.2
toml==0.10.2
openai==2.30.0
psycopg2-binary==2.9.11
pydantic==2.12.5